fix(core_ui): relocate AiroFold from app/ to core_ui
feature_iptv (framework layer) needs AiroFold for its player layout
decision, but app/ is the application layer that depends on feature_iptv —
not the reverse. Move the helper into core_ui, which feature_iptv already
depends on, instead of creating a layering violation.
